fix: fix definition files so nativescript-cloud can be successfully built
Currently `nativescript-cloud` cannot be built with latest master branch of CLI due to the following errors:
```
node_modules/nativescript/lib/definitions/project.d.ts(476,96): error TS2503: Cannot find namespace 'IXcode'.
node_modules/nativescript/lib/definitions/project.d.ts(476,165): error TS2503: Cannot find namespace 'IXcode'.
node_modules/nativescript/lib/definitions/project.d.ts(479,128): error TS2503: Cannot find namespace 'IXcode'.
node_modules/nativescript/lib/definitions/project.d.ts(480,99): error TS2503: Cannot find namespace 'IXcode'.
```

The issue happens as `IXcode` interface is already used from project.d.ts file. As `nativescript-cloud` lib generates `references.d.ts` file with references to all `.d.ts` files from NativeScript CLI, we need a separate `.d.ts`  file for `IXcode` interface. This way, it will be included in the generated `references.d.ts` file and cloud-lib will be successfully built.
