+++ lib/node/xcode.ts
@@ -1,14 +1,5 @@
 import * as xcode from "nativescript-dev-xcode";
 
-declare global {
-	type IXcode = typeof xcode;
-	export namespace IXcode {
-		export type target = xcode.target;
-		export type project = xcode.project;
-		export interface Options extends xcode.Options {} // tslint:disable-line
-	}
-}
-
 export { xcode };
 
 $injector.register("xcode", xcode);
